STCOR-789: restore original url after login
Refs STCOR-789
