+++ src/components/OIDCRedirect.js
@@ -1,17 +1,24 @@
 import { withRouter, Redirect, useLocation } from 'react-router';
 import queryString from 'query-string';
+import { useStripes } from '../StripesContext';
 
 /**
  * OIDCRedirect authenticated route handler for /oidc-landing.
  *
+ * Read unauthorized_path from session storage if keycloak authn provided
+ * if `fwd` provided into redirect url, it causes strange behavior - infinite login requests
+ * decided to use session storage for that case. Refs STCOR-789
+ *
  * Reads `fwd` from URL params and redirects.
  *
  * @see RootWithIntl
+ * @see AuthnLogin
  *
  * @returns {Redirect}
  */
 const OIDCRedirect = () => {
   const location = useLocation();
+  const stripes = useStripes();
 
   const getParams = () => {
     const search = location.search;
@@ -20,6 +27,11 @@ const OIDCRedirect = () => {
   };
 
   const getUrl = () => {
+    if (stripes.okapi.authnUrl) {
+      const unauthorizedPath = sessionStorage.getItem('unauthorized_path');
+      if (unauthorizedPath) return unauthorizedPath;
+    }
+
     const params = getParams();
     return params?.fwd ?? '';
   };
